Remove dependence on colors 16-21 for goog dark
This removes the dependency on colors 16-21. This way, when a terminal
does not support overriding ANSI color codes, at least we don't end up
with unusable garbage with blue colors everywhere inside VIM.

This is part of a personal effort to remove all dependencies on the
ability of overriding ANSI colors. I don't explicitly override colors
0-15 though, because I do like the google-dark color scheme, so if
terminals do have the capability to override ANSI codes, at least I'll
participate in that.